first iteration of comment handling

+++ backend/app.py
@@ -342,7 +342,7 @@ def logout_user():
         token_value = headers["Authorization"]
         token = Token.query.get(token_value)
 
-        if token_value == token.token:
+        if token and token_value == token.token:
             # user_logout_message = "{0} is now logged out on this device".format(str(g.user))
             g.user.tokens.remove(token)
             db.session.delete(token)
@@ -626,6 +626,7 @@ def get_match_data(match):
         comment_data = {'id': comment.id,
                         'author': comment.author_id,
                         'message': comment.message,
+                        'name': User.query.get(comment.author_id).name,
                         'date': comment.date}
 
         comment_list.append(comment_data)
